fix: adjust logic of components
