+++ src/pages/moneyDestination/tabs/moneyDestinationTab/AddDestination/index.tsx
@@ -1,62 +1,25 @@
-import { useEffect, useRef, useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { FormikProps } from "formik";
-import { IRuleDecision } from "@isettingkit/input";
-
 import { addDestinationStepsConfig } from "./config/assisted.config";
 import { AddDestinationUI } from "./interface";
-import { IGeneralInformationEntry } from "../forms/GeneralInformation/types";
+
+import { useAddDestination } from "@hooks/MoneyDestination/useAddDestination";
 
 function AddDestination() {
-  const [currentStep, setCurrentStep] = useState(1);
-  const [formValues, setFormValues] = useState<IGeneralInformationEntry>({
-    nameDestination: "",
-    description: "",
-  });
-
-  const [showModal, setShowModal] = useState(false);
-  const [isCurrentFormValid, setIsCurrentFormValid] = useState(false);
-  const [creditLineDecisions, setCreditLineDecisions] = useState<
-    IRuleDecision[]
-  >([]);
-
-  const navigate = useNavigate();
-
-  const generalInformationRef =
-    useRef<FormikProps<IGeneralInformationEntry>>(null);
-
-  const [nameDecision, setNameDecision] = useState(
-    generalInformationRef.current?.values.nameDestination ?? "",
-  );
-
-  useEffect(() => {
-    setNameDecision(formValues.nameDestination ?? "");
-  }, [formValues.nameDestination]);
-
-  const handleNextStep = () => {
-    if (currentStep < addDestinationStepsConfig("").length) {
-      if (generalInformationRef.current) {
-        setFormValues(generalInformationRef.current.values);
-        setIsCurrentFormValid(generalInformationRef.current.isValid);
-      }
-      setCurrentStep(currentStep + 1);
-    }
-  };
-
-  const handlePreviousStep = () => {
-    if (currentStep > 1) {
-      setCurrentStep(currentStep - 1);
-    }
-  };
-
-  const handleToggleModal = () => {
-    setShowModal(!showModal);
-  };
-
-  const handleSubmitClick = () => {
-    handleToggleModal();
-    navigate("/money-destination");
-  };
+  const {
+    creditLineDecisions,
+    currentStep,
+    formValues,
+    generalInformationRef,
+    isCurrentFormValid,
+    nameDecision,
+    showModal,
+    handleNextStep,
+    handlePreviousStep,
+    handleSubmitClick,
+    handleToggleModal,
+    setCreditLineDecisions,
+    setCurrentStep,
+    setIsCurrentFormValid,
+  } = useAddDestination();
 
   return (
     <AddDestinationUI
