+++ src/pages/moneyDestination/tabs/moneyDestinationTab/forms/GeneralInformation/index.tsx
@@ -1,35 +1,11 @@
-import {
-  forwardRef,
-  useContext,
-  useEffect,
-  useImperativeHandle,
-  useState,
-} from "react";
-import { FormikProps, useFormik } from "formik";
-import { object } from "yup";
+import { forwardRef, useContext } from "react";
+import { FormikProps } from "formik";
 
-import { validationMessages } from "@validations/validationMessages";
 import { AppContext } from "@context/AppContext";
-import { validationRules } from "@validations/validationRules";
 import { useEnumMoneyDestination } from "@hooks/MoneyDestination/useEnumMoneyDestination";
-import { IServerDomain } from "@ptypes/domain.types";
-import {
-  normalizeCodeDestination,
-  normalizeDestination,
-  normalizeNameDestination,
-} from "@utils/destination";
+import { useGeneralInformationForm } from "@hooks/MoneyDestination/useGeneralInformationForm";
 import { GeneralInformationFormUI } from "./interface";
 import { IGeneralInformationEntry } from "./types";
-
-const createValidationSchema = () =>
-  object().shape({
-    nameDestination: validationRules.string.required(
-      validationMessages.required,
-    ),
-    description: validationRules.string.required(validationMessages.required),
-  });
-
-const validationSchema = createValidationSchema();
 
 interface IGeneralInformationForm {
   initialValues: IGeneralInformationEntry;
@@ -43,63 +19,21 @@ const GeneralInformationForm = forwardRef<
   FormikProps<IGeneralInformationEntry>,
   IGeneralInformationForm
 >(({ initialValues, onFormValid, onSubmit, handleNextStep, loading }, ref) => {
-  const formik = useFormik({
-    initialValues,
-    validationSchema,
-    validateOnBlur: false,
-    onSubmit: onSubmit ?? (() => true),
-  });
-
   const { appData } = useContext(AppContext);
-  const [autosuggestValue, setAutosuggestValue] = useState(
-    formik.values.nameDestination || "",
-  );
 
   const { enumData } = useEnumMoneyDestination(
     "moneydestination",
     appData.businessUnit.publicCode,
   );
 
-  const optionsDestination: IServerDomain[] = enumData.map((item) => ({
-    id: item.code,
-    label: normalizeNameDestination(item.code)?.name as unknown as string,
-    value: normalizeNameDestination(item.code)?.name as unknown as string,
-  }));
-
-  useImperativeHandle(ref, () => formik);
-
-  useEffect(() => {
-    if (onFormValid) {
-      formik.validateForm().then((errors) => {
-        const isFormValid = Object.keys(errors).length === 0;
-        onFormValid(isFormValid);
-      });
-    }
-  }, [formik.values, onFormValid]);
-
-  useEffect(() => {
-    setAutosuggestValue(formik.values.nameDestination || "");
-  }, [formik.values.nameDestination]);
-
-  const handleChange = (name: string, value: string) => {
-    setAutosuggestValue(value);
-    formik.setFieldValue(name, value);
-
-    if (name === "nameDestination") {
-      if (value === "") {
-        formik.setFieldValue("description", "");
-      } else {
-        const nameEnum = normalizeCodeDestination(value)?.code || "";
-        const description = normalizeDestination(
-          enumData,
-          nameEnum,
-        )?.description;
-        if (description !== undefined) {
-          formik.setFieldValue("description", description);
-        }
-      }
-    }
-  };
+  const { autosuggestValue, optionsDestination, formik, handleChange } =
+    useGeneralInformationForm(
+      enumData,
+      initialValues,
+      ref,
+      onSubmit,
+      onFormValid,
+    );
 
   return (
     <GeneralInformationFormUI
