+++ src/pages/moneyDestination/tabs/moneyDestinationTab/forms/VerificationForm/types.ts
@@ -1,7 +1,9 @@
 import { IRuleDecision } from "@isettingkit/input";
 import { IGeneralInformationEntry } from "../GeneralInformation/types";
 
-export interface IFormsUpdateData {
+interface IFormsUpdateData {
   personalInformation: { isValid: boolean; values: IGeneralInformationEntry };
   creditline: { isValid: boolean; values: IRuleDecision[] };
 }
+
+export type { IFormsUpdateData };
